// apps/web/src/lib/api/generated/main/model/create-notification-rule-dto-schedule-anchor.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type CreateNotificationRuleDtoScheduleAnchor = typeof CreateNotificationRuleDtoScheduleAnchor[keyof typeof CreateNotificationRuleDtoScheduleAnchor];


export const CreateNotificationRuleDtoScheduleAnchor = {
  MIN_SPAWN: 'MIN_SPAWN',
  MAX_SPAWN: 'MAX_SPAWN',
} as const;