// apps/web/src/lib/api/generated/main/model/create-notification-rule-dto-schedule-interval-type.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type CreateNotificationRuleDtoScheduleIntervalType = typeof CreateNotificationRuleDtoScheduleIntervalType[keyof typeof CreateNotificationRuleDtoScheduleIntervalType];


export const CreateNotificationRuleDtoScheduleIntervalType = {
  ONCE: 'ONCE',
  HOURLY: 'HOURLY',
  DAILY: 'DAILY',
  WEEKLY: 'WEEKLY',
} as const;