// apps/web/src/lib/api/generated/main/model/createNotificationRuleDtoScheduleStrategy.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export type CreateNotificationRuleDtoScheduleStrategy = typeof CreateNotificationRuleDtoScheduleStrategy[keyof typeof CreateNotificationRuleDtoScheduleStrategy];


export const CreateNotificationRuleDtoScheduleStrategy = {
  SPAWN_WINDOW_RELATIVE: 'SPAWN_WINDOW_RELATIVE',
  FIXED_DATETIME: 'FIXED_DATETIME',
} as const;